Clean the WpsService.
