+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/ui/dataui/DataFieldUI.java
@@ -23,7 +23,7 @@ import net.miginfocom.swing.MigLayout;
 import org.orbisgis.sif.common.ContainerItem;
 import org.orbisgis.wpsclient.WpsClient;
 import org.orbisgis.wpsclient.view.utils.ToolBoxIcon;
-import org.orbisgis.wpsservice.WpsService;
+import org.orbisgis.wpsservice.WpsServiceImplementation;
 import org.orbisgis.wpsservice.model.*;
 
 import javax.swing.*;
@@ -136,7 +136,7 @@ public class DataFieldUI implements DataUI{
             dataField.setSourceModified(false);
             String tableName = ((URI) dataMap.get(dataField.getDataStoreIdentifier())).getFragment();
             comboBox.removeAllItems();
-            for (String field : WpsService.getTableFieldList(tableName, dataField.getFieldTypeList())) {
+            for (String field : WpsServiceImplementation.getTableFieldList(tableName, dataField.getFieldTypeList())) {
                 comboBox.addItem(field);
             }
             if(isOptional) {
