+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/utils/ExecutionWorker.java
@@ -21,7 +21,7 @@ package org.orbisgis.wpsclient.view.utils;
 
 import org.orbisgis.commons.progress.SwingWorkerPM;
 import org.orbisgis.wpsclient.view.utils.editor.process.ProcessEditableElement;
-import org.orbisgis.wpsservice.WpsService;
+import org.orbisgis.wpsservice.WpsServiceImplementation;
 import org.orbisgis.wpsservice.model.Process;
 
 import java.net.URI;
@@ -47,7 +47,7 @@ public class ExecutionWorker extends SwingWorkerPM{
     /** Input and output data map */
     private Map<URI, Object> dataMap;
     /** ToolBox */
-    private WpsService wpsService;
+    private WpsServiceImplementation wpsServiceImplementation;
     /** Process element containing all the information it */
     private ProcessEditableElement pee;
 
@@ -55,18 +55,18 @@ public class ExecutionWorker extends SwingWorkerPM{
      * Main constructor.
      * @param pee ProcessEditableElement which will be used to communicate the state and the log of the process.
      */
-    public ExecutionWorker(ProcessEditableElement pee, WpsService wpsService){
+    public ExecutionWorker(ProcessEditableElement pee, WpsServiceImplementation wpsServiceImplementation){
         this.pee = pee;
         this.process = pee.getProcess();
         this.dataMap = new HashMap<>();
         this.dataMap.putAll(pee.getInputDataMap());
         this.dataMap.putAll(pee.getOutputDataMap());
-        this.wpsService = wpsService;
+        this.wpsServiceImplementation = wpsServiceImplementation;
     }
 
     @Override
     protected Object doInBackground() throws Exception {
-        wpsService.executeProcess(process, dataMap, pee);
+        wpsServiceImplementation.execute(process, dataMap, pee);
         return null;
     }
 }
