+++ bundles/wps-client/src/main/java/org/orbisgis/wpsclient/view/utils/editor/process/ProcessEditor.java
@@ -166,7 +166,7 @@ public class ProcessEditor extends JPanel implements EditorDockable, PropertyCha
         wpsClient.validateInstance(this);
         pee.setProcessState(ProcessEditableElement.ProcessState.RUNNING);
         //Run the process in a separated thread
-        thread = new ExecutionWorker(pee, wpsClient.getWpsService());
+        thread = new ExecutionWorker(pee, wpsClient.getWpsServiceImplementation());
         wpsClient.getExecutorService().execute(thread);
         return false;
     }
