+++ bundles/wps-service/src/main/java/org/orbisgis/wpsservice/controller/execution/DataProcessing.java
@@ -19,7 +19,7 @@
 
 package org.orbisgis.wpsservice.controller.execution;
 
-import org.orbisgis.wpsservice.WpsService;
+import org.orbisgis.wpsservice.WpsServiceImplementation;
 import org.orbisgis.wpsservice.model.DataDescription;
 import org.orbisgis.wpsservice.model.DescriptionType;
 
@@ -34,6 +34,6 @@ import java.util.Map;
 public interface DataProcessing {
 
     Class<? extends DataDescription> getDataClass();
-    Map<URI, Object> preProcessData(WpsService wpsService, DescriptionType inputOrOutput, Map<URI, Object> dataMap);
-    void postProcessData(WpsService wpsService, DescriptionType inputOrOutput, Map<URI, Object> dataMap, Map<URI, Object> stash);
+    Map<URI, Object> preProcessData(WpsServiceImplementation wpsServiceImplementation, DescriptionType inputOrOutput, Map<URI, Object> dataMap);
+    void postProcessData(WpsServiceImplementation wpsServiceImplementation, DescriptionType inputOrOutput, Map<URI, Object> dataMap, Map<URI, Object> stash);
 }
